# palabatu-be/docs/swagger.json
{
    "components": {
        "schemas": {
...
            "internal_crags.Crag": {
                "properties": {
                    "access_notes": {
                        "type": "string"
                    },
                    "created_at": {
                        "type": "string"
                    },
                    "created_by": {
                        "type": "string"
                    },
                    "directions": {
                        "type": "string"
                    },
                    "id": {
                        "type": "string"
                    },
                    "lat": {
                        "type": "number"
                    },
                    "lng": {
                        "type": "number"
                    },
                    "name": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "internal_crags.CragListItem": {
                "properties": {
                    "access_notes": {
                        "type": "string"
                    },
                    "boulder_count": {
                        "type": "integer"
                    },
                    "created_at": {
                        "type": "string"
                    },
                    "created_by": {
                        "type": "string"
                    },
                    "creator_name": {
                        "type": "string"
                    },
                    "directions": {
                        "type": "string"
                    },
                    "id": {
                        "type": "string"
                    },
                    "lat": {
                        "type": "number"
                    },
                    "lng": {
                        "type": "number"
                    },
                    "name": {
                        "type": "string"
                    },
                    "problem_count": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
            "internal_crags.CreateCragRequest": {
                "properties": {
                    "access_notes": {
                        "type": "string"
                    },
                    "directions": {
                        "type": "string"
                    },
                    "lat": {
                        "type": "number"
                    },
                    "lng": {
                        "type": "number"
                    },
                    "name": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "internal_crags.UpdateCragRequest": {
                "properties": {
                    "access_notes": {
                        "type": "string"
                    },
                    "directions": {
                        "type": "string"
                    },
                    "lat": {
                        "type": "number"
                    },
                    "lng": {
                        "type": "number"
                    },
                    "name": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "palabatu-be_internal_apitypes.ErrorResponse": {
                "properties": {
                    "error": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "BearerAuth": {
                "description": "Type \"Bearer\" followed by a space and the JWT, e.g. \"Bearer eyJhbGc...\".",
                "in": "header",
                "name": "Authorization",
                "type": "apiKey"
            }
        }
    },
    "info": {
        "description": "Backend API for Palabatu, a community app for Indonesian bouldering enthusiasts (spot map, climber profiles, route/problem listings, comments, reactions, reports).",
        "title": "Palabatu API",
        "version": "1.0"
    },
...
    "paths": {
...
        "/api/crags": {
            "get": {
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/internal_crags.CragListItem"
                                    },
                                    "type": "array"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/palabatu-be_internal_apitypes.ErrorResponse"
                                }
                            }
                        },
                        "description": "Internal Server Error"
                    }
                },
                "summary": "List all crags",
                "tags": [
                    "crags"
                ]
            },
            "post": {
                "description": "Any authenticated user may create a crag; no role gate.",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "oneOf": [
                                    {
                                        "type": "object"
                                    },
                                    {
                                        "$ref": "#/components/schemas/internal_crags.CreateCragRequest",
                                        "summary": "body",
                                        "description": "New crag"
                                    }
                                ]
                            }
                        }
                    },
                    "description": "New crag",
                    "required": true
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/internal_crags.Crag"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/palabatu-be_internal_apitypes.ErrorResponse"
                                }
                            }
                        },
                        "description": "Bad Request"
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/palabatu-be_internal_apitypes.ErrorResponse"
                                }
                            }
                        },
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "Create a crag",
                "tags": [
                    "crags"
                ]
            }
        },
        "/api/crags/{id}": {
            "get": {
                "parameters": [
                    {
                        "description": "Crag ID",
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/internal_crags.CragListItem"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "404": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/palabatu-be_internal_apitypes.ErrorResponse"
                                }
                            }
                        },
                        "description": "Not Found"
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/palabatu-be_internal_apitypes.ErrorResponse"
                                }
                            }
                        },
                        "description": "Internal Server Error"
                    }
                },
                "summary": "Get a crag",
                "tags": [
                    "crags"
                ]
            },
            "put": {
                "description": "Allowed for admins (Council/Associate title) on any crag, or the crag's own creator.",
                "parameters": [
                    {
                        "description": "Crag ID",
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "oneOf": [
                                    {
                                        "type": "object"
                                    },
                                    {
                                        "$ref": "#/components/schemas/internal_crags.UpdateCragRequest",
                                        "summary": "body",
                                        "description": "Updated fields"
                                    }
                                ]
                            }
                        }
                    },
                    "description": "Updated fields",
                    "required": true
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/internal_crags.Crag"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/palabatu-be_internal_apitypes.ErrorResponse"
                                }
                            }
                        },
                        "description": "Bad Request"
                    },
                    "403": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/palabatu-be_internal_apitypes.ErrorResponse"
                                }
                            }
                        },
                        "description": "not the creator and not an admin"
                    },
                    "404": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/palabatu-be_internal_apitypes.ErrorResponse"
                                }
                            }
                        },
                        "description": "Not Found"
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/palabatu-be_internal_apitypes.ErrorResponse"
                                }
                            }
                        },
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "Update a crag",
                "tags": [
                    "crags"
                ]
            }
        },
...
    },
    "openapi": "3.1.0",
    "servers": [
        {
            "url": "localhost:3001/"
        }
    ]
}